poc-1-pipeline-prototype-bruteforce-data/requirements.txt
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.2
fairlearn==0.10.0
matplotlib==3.7.1
openai
python-dotenv
aif360==0.5.0
faker

# just to get rid of the warnings...
aif360[AdversarialDebiasing] 
aif360[LawSchoolGPA]
tempeh